requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
kiteconnect>=5.0.0
yfinance>=0.2.40
pandas>=2.0.0
pandas-ta>=0.3.14b0
numpy>=1.26.0
plotly>=5.20.0
scipy>=1.12.0
feedparser>=6.0.11
python-dotenv>=1.0.0